+++ packages/yoga/src/Theme/themes/BaseTheme.js
@@ -2,6 +2,7 @@ import tokens from '@gympass/yoga-tokens';
 import { hexToRgb } from '@gympass/yoga-common';
 
 const {
+  breakpoints,
   spacing,
   radii,
   fontSizes,
@@ -398,6 +399,25 @@ const BaseTheme = ({ primary, secondary, tertiary }) => {
         },
       },
     },
+    grid: {
+      gutter: {
+        desktop: 24,
+        mobile: 16,
+      },
+      container: {
+        width: 1600,
+        margin: {
+          desktop: {
+            left: 71,
+            right: 71,
+          },
+          mobile: {
+            left: 20,
+            right: 20,
+          },
+        },
+      },
+    },
     text: {
       h1: {
         fontsize: fontSizes.xxlarge,
@@ -437,7 +457,15 @@ const BaseTheme = ({ primary, secondary, tertiary }) => {
     },
   };
 
-  return { components, baseFontFamily, baseFontSize, colors, spacing, fonts };
+  return {
+    components,
+    baseFontFamily,
+    baseFontSize,
+    colors,
+    spacing,
+    fonts,
+    breakpoints,
+  };
 };
 
 export default BaseTheme;